rewrite deduplication tests
